requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.61.0
anyio==4.10.0
argcomplete==3.6.2
attrs==25.3.0
Authlib==1.6.1
boto3==1.40.3
botocore==1.40.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cohere==5.16.2
colorama==0.4.6
cryptography==45.0.6
cyclopts==3.22.5
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastavro==1.12.0
fastmcp==2.11.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.28.0
gotrue==2.12.3
griffe==1.10.0
groq==0.31.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.7
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
logfire-api==4.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.3
mdurl==0.1.2
mistralai==1.9.3
more-itertools==10.7.0
multidict==6.6.3
openai==1.99.1
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.36.0
packaging==25.0
parse==1.20.2
pathable==0.4.4
postgrest==1.1.1
prompt_toolkit==3.0.51
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-ai==0.5.1
pydantic-ai-slim==0.5.1
pydantic-evals==0.5.1
pydantic-graph==0.5.1
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.7.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.0
supafunc==0.10.1
tenacity==8.5.0
tokenizers==0.21.4
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uv==0.8.5
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.1
yarl==1.20.1
zipp==3.23.0
